+++ atl-checker/src/game_structure/lcgs/mod.rs
