+++ atl-checker/src/game_structure/mod.rs
